requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
chromadb
python-dotenv
twilio
pydantic
